// src/app/models/noconformidad.model.ts
export class NoConformidadModel {

  id: any;
  fecha_evidencia: any;
  proceso: String;
  tipo_accion: String;
  descripcion: String;
  tipo_grado: String;
  quien_detecta: String;
  numeral_norma: String;
  resolver: any;
  analisis: String;
  accion_correctiva: String;
  verificacion: String;
  fecha_verificacion: any;
  estado: String;
  usuario: any;
  obra: any;

}